Atualizacao versao CRUD

+++ models/Pedido.js
@@ -1,18 +1,18 @@
-import Sequelize from "sequelize";
-import connection from "../config/sequelize-config.js";
+    import Sequelize from "sequelize";
+    import connection from "../config/sequelize-config.js";
 
-const Pedido = connection.define('pedidos', {
-    numero: {
-        type: Sequelize.BIGINT,
-        allowNull: false,
-        autoIncrement: true,
-        primaryKey: true,
-    },
-    valor: {
-        type: Sequelize.DOUBLE,
-        allowNull: false,
-    },
-});
+    const Pedido = connection.define('pedidos', {
+        numero: {
+            type: Sequelize.INTEGER,
+            allowNull: false,
+            autoIncrement: true,
+            primaryKey: true,
+        },
+        valor: {
+            type: Sequelize.FLOAT,
+            allowNull: false,
+        },
+    });
 
-Pedido.sync({ force: false }); 
-export default Pedido;
+    Pedido.sync({ force: false }); 
+    export default Pedido;
